indexer/requirements.txt
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
sentence-transformers==2.6.1
qdrant-client==1.9.0
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
requests==2.32.3
pillow==10.4.0
pymupdf==1.24.7
pytesseract==0.3.10
numpy==1.26.4